add ImageWithCollapse to support better accessibilty for images with text
