+++ pages/projects/dashboard/index.js
@@ -7,6 +7,7 @@ import { ProjectInfo } from "../../../components/atoms/ProjectInfo";
 import { createBreadcrumbs } from "../../../lib/utils/createBreadcrumbs";
 import { Heading } from "../../../components/molecules/Heading";
 import { Collapse } from "../../../components/molecules/Collapse";
+import { generateCollapseElements } from "../../../lib/utils/generateCollapseElements";
 
 export default function MscaDashboard(props) {
   const pageData = props.pageData?.item;
@@ -29,50 +30,6 @@ export default function MscaDashboard(props) {
       "gc:custom/decd-endc/project-stage/beta": "Bêta",
     },
   };
-
-  function generateReactElements(json) {
-    const elements = [];
-
-    for (const item of json) {
-      if (item.nodeType === "paragraph") {
-        elements.push(
-          <p key={elements.length}>{generateReactElements(item.content)}</p>
-        );
-      } else if (item.nodeType === "unordered-list") {
-        const listItems = item.content.map((listItem, index) => (
-          <li key={index} className="my-0">
-            {generateReactElements(listItem.content)}
-          </li>
-        ));
-        elements.push(
-          <ul key={elements.length} className="mb-0">
-            {listItems}
-          </ul>
-        );
-      } else if (item.nodeType === "ordered-list") {
-        const listItems = item.content.map((listItem, index) => (
-          <li key={index} className="my-0">
-            {generateReactElements(listItem.content)}
-          </li>
-        ));
-        elements.push(
-          <ol key={elements.length} className="mb-0">
-            {listItems}
-          </ol>
-        );
-      } else if (item.nodeType === "list-item") {
-        elements.push(
-          <li key={elements.length} className="my-0">
-            {generateReactElements(item.content)}
-          </li>
-        );
-      } else if (item.nodeType === "text") {
-        elements.push(item.value);
-      }
-    }
-
-    return elements;
-  }
 
   useEffect(() => {
     if (props.adobeAnalyticsUrl) {
@@ -486,7 +443,7 @@ export default function MscaDashboard(props) {
                           : pageData.scFragments[4].scFragments[0]
                               .scFragments[0].scLongDescHeadingFr
                       }
-                      children={generateReactElements(
+                      children={generateCollapseElements(
                         props.locale === "en"
                           ? pageData.scFragments[4].scFragments[0]
                               .scFragments[0].scLongDescEn.json
@@ -560,7 +517,7 @@ export default function MscaDashboard(props) {
                           : pageData.scFragments[4].scFragments[1]
                               .scFragments[0].scLongDescHeadingFr
                       }
-                      children={generateReactElements(
+                      children={generateCollapseElements(
                         props.locale === "en"
                           ? pageData.scFragments[4].scFragments[1]
                               .scFragments[0].scLongDescEn.json
@@ -662,7 +619,7 @@ export default function MscaDashboard(props) {
                           : pageData.scFragments[4].scFragments[2]
                               .scFragments[0].scLongDescHeadingFr
                       }
-                      children={generateReactElements(
+                      children={generateCollapseElements(
                         props.locale === "en"
                           ? pageData.scFragments[4].scFragments[2]
                               .scFragments[0].scLongDescEn.json
@@ -750,7 +707,7 @@ export default function MscaDashboard(props) {
                           : pageData.scFragments[4].scFragments[3]
                               .scFragments[0].scLongDescHeadingFr
                       }
-                      children={generateReactElements(
+                      children={generateCollapseElements(
                         props.locale === "en"
                           ? pageData.scFragments[4].scFragments[3]
                               .scFragments[0].scLongDescEn.json
